// src/presentation/enum/presentation.enum.ts
export enum PresPromptType {
  PRESENTATION_NOTE = 'PRESENTATION_NOTE',
  PRESENTATION_SLIDE = 'PRESENTATION_SLIDE',
  PRESENTATION_SCRIPT = 'PRESENTATION_SCRIPT',
  PRESENTATION_OUTLINE = 'PRESENTATION_OUTLINE',
}

export enum SlideAssPromptType {
  PRESENTATION_TITLE = 'PRESENTATION_TITLE',
  PRESENTATION_SUBTITLE = 'PRESENTATION_SUBTITLE',
  PRESENTATION_AUDIENCE = 'PRESENTATION_AUDIENCE',
  PRESENTATION_GOAL = 'PRESENTATION_GOAL',
  PRESENTATION_OUTLINE = 'PRESENTATION_OUTLINE',
  PRESENTATION_DESCRIPTION = 'PRESENTATION_DESCRIPTION',
}

export enum AccessModeType {
  PUBLIC = 'PUBLIC',
  PRIVATE = 'PRIVATE',
}

export enum QuizOrPollType {
  QUIZ = 'QUIZ',
  POLL = 'POLL',
}